Fix date cast
