+++ src/main/kotlin/com/projectcitybuild/entities/LegacyWarp.kt
@@ -3,7 +3,6 @@ package com.projectcitybuild.entities
 import com.projectcitybuild.entities.serializables.SerializableDate
 import com.projectcitybuild.entities.serializables.SerializableUUID
 import kotlinx.serialization.Serializable
-import java.sql.Date
 
 @Serializable
 data class LegacyWarp(
@@ -18,14 +17,3 @@ data class LegacyWarp(
     val createdAt: SerializableDate,
 )
 
-data class Warp(
-    val name: String,
-    val serverName: String,
-    val worldName: String,
-    val x: Double,
-    val y: Double,
-    val z: Double,
-    val pitch: Float,
-    val yaw: Float,
-    val createdAt: Date,
-)
